DOC-9474 allow connection to Capella
wait-for-couchbase.sh still needs work
(which means the automated Docker option won't work.)

But documented for the "Running the application manually"

This commit adds the `dnssrv` and `cert` properties
which are required to connect to a Capella instance.

DOC-9474 kludge POC to connect to Capella

This requires hard-coding credentials in Database.java,
as our instructions for `mvn spring-boot:run ...` don't
seem to pass these in properly.

Future TODO: made the SRV/cert/(or no-verify?) configurable.

wip